Convert GLTF Document to use EditorSceneImporterMeshNode3D.
GLTF should work in a game template.

Add a GLTFDocument extension list and an extension for converting mesh instance 3d to the editor mesh node.

Use GLTF module when the editor tools are disabled.

Modified the render server to be less restrictive on matching blend arrays and have more logging.