openxr: call xrBeginFrame even if shouldRender == true
```
As long as the session is running, the application should keep running
the frame loop to maintain the frame synchronization to the runtime, even
if this requires calling xrEndFrame with all layers omitted.
```

In OpenXRDevice::end_frame() we return early after calling xrEndFrame with
0 layers submitted if shouldRender is false.
This early return still requires us to have called xrBeginFrame.